requirements.txt
# ---- Core Django & DRF ----
Django>=5.0,<6.0
djangorestframework>=3.15
drf-spectacular>=0.27.0

# ---- Database Driver ----
psycopg2-binary>=2.9

# ---- Development Utilities ----
python-dotenv>=1.0
gunicorn>=21.2

# ---- Optional but useful ----
django-cors-headers>=4.4
